refactor/87: rename Board components for naming convention
