+++ src/components/About/About.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallowMatchSnapshot } from '../../common/test_utils';
-import About from './About';
+import About from './About.component';
 
 describe('About tests', () => {
   test('default renderer', () => {
