+++ src/components/Board/Board.component.js
@@ -65,7 +65,7 @@ export class Board extends Component {
      */
     onRequestPreviousBoard: PropTypes.func,
     /**
-     * Callback fired when a board button is focuesd
+     * Callback fired when a board button is focused
      */
     onFocusBoardButton: PropTypes.func,
     /**
