+++ src/components/Board/BoardButton/index.js
@@ -1,1 +1,1 @@
-export { default } from './BoardButton';
+export { default } from './BoardButton.component';
