+++ src/components/Board/Navbar/index.js
@@ -1,1 +1,1 @@
-export { default } from './Navbar';
+export { default } from './Navbar.component';
