+++ src/components/Board/SymbolOutput/SymbolOutput.component.js
